+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `--content-type` option for requests with multiple allowed content types. (CLI)
+
 ### Changed
 
 - Moved from net7 to net8.